Permissions fixed for face recog. DB and its dir
The face recognition database and its directory
are protected through the usage of the AEGIS
"powered" gallerycoredata-user user and
gallerycoredata-users group.

Also, the file permisssions mask for them were 070
in the case of the directory and 060 in the case
of the database.

The database is in SQLite format. When doing
transactions to it, the SQLite driver creates
temporal files as the journal one, to be able to
recover the database under disaster.

This journal file gets the UID and GID of the
running process and the permissions from a
combination of the permissions of the original
database file and the running process' umask.

As a consequence, the journal file usually has the
permissions mask 040.

This is a problem when a disaster happens in the
database since, although we hold the journaling
file, the owner of that file is not able to read
it and, hence, the SQLite database remains useless
for the processes with the same UID than the owner
of the journal file, even when they belong to the
same group than that file.

From now on, when selecting to protect, we will
set the SGID and the 070 mask to the DB directory.
In the case of the files, we will set the 660 mask
to them.

Modified the protectDB function of the
GalleryPluginFacerecognitionResetterPrivate class
so it performs the operations described above.